feat: upgrade safe network to 0.28.x
The 0.28.x release of safe_network is quite a signifcant release, as it changes the client blob API
and also makes some changes to the way the client sets up a connection.

Due to the fact that the client that the client constructor now *requires* a bootstrap
configuration, the API was changed to require callers to pass the bootstrap config, rather than
using an `Option`. The type of the config was also changed again, from a `Vec` to a `BTreeSet`, and
this required another few changes in the API.

The functions for reading and writing blob data to the network were changed to work with the `bytes`
crate, rather than slices of bytes. We agreed to change the API to also replace its use of `&[u8]`
with `Bytes`, rather than do a conversion. Quite a few tests were subsequently updated for this
change.