feat: upgrade safe network to 0.31.x
Version 0.29.x onwards changes the way we connect to the network. What
we're now calling the 'node config' rather than 'bootstrap config',
has 2 components: a public key and a list of nodes, where the nodes are
an IP and port pair. The public key is a BLS key and what we refer to as
the 'genesis key'. It's a key that's generated when the first node on
the network is started.

For this reason, the BootstrapConfig type was renamed to NodeConfig and
changed to a tuple with a public key and a list of nodes.