fix: use correct data type on multimap entries
BREAKING CHANGE: Multimap entries will be assigned a different data type when they are created.

In the same way the xorurls for files and serialized files containers needed to have the correct
data type specified as a spot or a blob, so too did entries for multimaps that get used as part of
the NRS system.

As per files, if the multimap entry is less than 3072 bytes, it'll be stored as a spot, otherwise,
it'll be stored as a blob.